requirements.txt
SQLAlchemy==1.4.31
PyYAML==6.0
psycopg2-binary==2.9.1
